Improve calculations with nested dice groups
A dice group will be treated as a discrete unit if and only if it is
enclosed in parens. That is: `(1d2 + 1d4 + 2d6)h` will roll four dice
and take the highest single dice of the four. But `(1d2 + 1d4 + (2d6))h`
will roll four dice but take the highest of the d2, the d4, or the sum
of both d6.